requirements.txt
undetected-chromedriver>=3.5.5
selenium>=4.15.0
requests>=2.31.0
PyYAML>=6.0
